# Conflicts:
#	Character/PlayerCharacter.cpp
#	Character/PlayerCharacter.hpp
#	Core/Tile.cpp
#	Core/Tile.hpp
#	main.cpp

+++ Character/ResourceBar.cpp
@@ -69,7 +69,7 @@ void ResourceBar::setSize(const sf::Vector2f & newSize) {
 
 
 void ResourceBar::update() {
-	int unitsPerPixel = static_cast<int> (size.x / maxValue);
+	float unitsPerPixel = size.x / maxValue;
 	barInnerMeter.setSize(sf::Vector2f(static_cast<float>(currentValue * unitsPerPixel), size.y));
 }
